requirements.txt
fastapi==0.99.0
uvicorn==0.23.2
pydantic[email]==1.10.15
beanie==1.24.0
PyJWT==2.8.0
aiohttp~=3.8.6
python-decouple==3.8
pytz~=2023.3.post1